Split up rep.rs
