+++ src/container.rs
@@ -11,7 +11,7 @@ use futures_util::{
 };
 use hyper::Body;
 use mime::Mime;
-use serde::Serialize;
+use serde::{Deserialize, Serialize};
 use serde_json::{json, Map, Value};
 use url::form_urlencoded;
 
@@ -19,12 +19,16 @@ use crate::{
     docker::Docker,
     errors::{Error, Result},
     exec::{Exec, ExecContainerOptions},
-    rep::{
-        Change, Container as ContainerInfo, ContainerCreateInfo, ContainerDetails, Exit, Stats, Top,
-    },
+    image::Config,
+    network::{NetworkInfo, NetworkSettings},
     transport::Payload,
     tty::{self, Multiplexer as TtyMultiPlexer},
 };
+
+#[cfg(feature = "chrono")]
+use crate::datetime::datetime_from_unix_timestamp;
+#[cfg(feature = "chrono")]
+use chrono::{DateTime, Utc};
 
 /// Interface for accessing and manipulating a docker container
 pub struct Container<'docker> {
@@ -1117,6 +1121,241 @@ impl RmContainerOptionsBuilder {
             params: self.params.clone(),
         }
     }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct ContainerInfo {
+    #[cfg(feature = "chrono")]
+    #[serde(deserialize_with = "datetime_from_unix_timestamp")]
+    pub created: DateTime<Utc>,
+    #[cfg(not(feature = "chrono"))]
+    pub created: u64,
+    pub command: String,
+    pub id: String,
+    pub image: String,
+    #[serde(rename = "ImageID")]
+    pub image_id: String,
+    pub labels: HashMap<String, String>,
+    pub names: Vec<String>,
+    pub ports: Vec<Port>,
+    pub state: String,
+    pub status: String,
+    pub size_rw: Option<u64>,
+    pub size_root_fs: Option<u64>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct ContainerDetails {
+    pub app_armor_profile: String,
+    pub args: Vec<String>,
+    pub config: Config,
+    #[cfg(feature = "chrono")]
+    pub created: DateTime<Utc>,
+    #[cfg(not(feature = "chrono"))]
+    pub created: String,
+    pub driver: String,
+    // pub ExecIDs: ??
+    pub host_config: HostConfig,
+    pub hostname_path: String,
+    pub hosts_path: String,
+    pub log_path: String,
+    pub id: String,
+    pub image: String,
+    pub mount_label: String,
+    pub name: String,
+    pub network_settings: NetworkSettings,
+    pub path: String,
+    pub process_label: String,
+    pub resolv_conf_path: String,
+    pub restart_count: u64,
+    pub state: State,
+    pub mounts: Vec<Mount>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct Mount {
+    pub source: String,
+    pub destination: String,
+    pub mode: String,
+    #[serde(rename = "RW")]
+    pub rw: bool,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct State {
+    pub error: String,
+    pub exit_code: u64,
+    #[cfg(feature = "chrono")]
+    pub finished_at: DateTime<Utc>,
+    #[cfg(not(feature = "chrono"))]
+    pub finished_at: String,
+    #[serde(rename = "OOMKilled")]
+    pub oom_killed: bool,
+    pub paused: bool,
+    pub pid: u64,
+    pub restarting: bool,
+    pub running: bool,
+    #[cfg(feature = "chrono")]
+    pub started_at: DateTime<Utc>,
+    #[cfg(not(feature = "chrono"))]
+    pub started_at: String,
+    pub status: String,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct HostConfig {
+    pub cgroup_parent: Option<String>,
+    #[serde(rename = "ContainerIDFile")]
+    pub container_id_file: String,
+    pub cpu_shares: Option<u64>,
+    pub cpuset_cpus: Option<String>,
+    pub memory: Option<u64>,
+    pub memory_swap: Option<i64>,
+    pub network_mode: String,
+    pub pid_mode: Option<String>,
+    pub port_bindings: Option<HashMap<String, Vec<HashMap<String, String>>>>,
+    pub privileged: bool,
+    pub publish_all_ports: bool,
+    pub readonly_rootfs: Option<bool>, /* pub RestartPolicy: ???
+                                        * pub SecurityOpt: Option<???>,
+                                        * pub Ulimits: Option<???>
+                                        * pub VolumesFrom: Option<??/> */
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct Port {
+    pub ip: Option<String>,
+    pub private_port: u64,
+    pub public_port: Option<u64>,
+    #[serde(rename = "Type")]
+    pub typ: String,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct Stats {
+    pub read: String,
+    pub networks: HashMap<String, NetworkInfo>,
+    pub memory_stats: MemoryStats,
+    pub blkio_stats: BlkioStats,
+    pub cpu_stats: CpuStats,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct MemoryStats {
+    pub max_usage: u64,
+    pub usage: u64,
+    pub failcnt: Option<u64>,
+    pub limit: u64,
+    pub stats: MemoryStat,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct MemoryStat {
+    pub total_pgmajfault: u64,
+    pub cache: u64,
+    pub mapped_file: u64,
+    pub total_inactive_file: u64,
+    pub pgpgout: u64,
+    pub rss: u64,
+    pub total_mapped_file: u64,
+    pub writeback: u64,
+    pub unevictable: u64,
+    pub pgpgin: u64,
+    pub total_unevictable: u64,
+    pub pgmajfault: u64,
+    pub total_rss: u64,
+    pub total_rss_huge: u64,
+    pub total_writeback: u64,
+    pub total_inactive_anon: u64,
+    pub rss_huge: u64,
+    pub hierarchical_memory_limit: u64,
+    pub hierarchical_memsw_limit: u64,
+    pub total_pgfault: u64,
+    pub total_active_file: u64,
+    pub active_anon: u64,
+    pub total_active_anon: u64,
+    pub total_pgpgout: u64,
+    pub total_cache: u64,
+    pub inactive_anon: u64,
+    pub active_file: u64,
+    pub pgfault: u64,
+    pub inactive_file: u64,
+    pub total_pgpgin: u64,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct CpuStats {
+    pub cpu_usage: CpuUsage,
+    pub system_cpu_usage: u64,
+    pub throttling_data: ThrottlingData,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct CpuUsage {
+    pub percpu_usage: Vec<u64>,
+    pub usage_in_usermode: u64,
+    pub total_usage: u64,
+    pub usage_in_kernelmode: u64,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct ThrottlingData {
+    pub periods: u64,
+    pub throttled_periods: u64,
+    pub throttled_time: u64,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct BlkioStats {
+    pub io_service_bytes_recursive: Vec<BlkioStat>,
+    pub io_serviced_recursive: Vec<BlkioStat>,
+    pub io_queue_recursive: Vec<BlkioStat>,
+    pub io_service_time_recursive: Vec<BlkioStat>,
+    pub io_wait_time_recursive: Vec<BlkioStat>,
+    pub io_merged_recursive: Vec<BlkioStat>,
+    pub io_time_recursive: Vec<BlkioStat>,
+    pub sectors_recursive: Vec<BlkioStat>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct BlkioStat {
+    pub major: u64,
+    pub minor: u64,
+    pub op: String,
+    pub value: u64,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct Change {
+    pub kind: u64,
+    pub path: String,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct Top {
+    pub titles: Vec<String>,
+    pub processes: Vec<Vec<String>>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct ContainerCreateInfo {
+    pub id: String,
+    pub warnings: Option<Vec<String>>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct Exit {
+    pub status_code: u64,
 }
 
 #[cfg(test)]
